Title: 길이 단위 환산하기, Time: 50ms, Memory: 16MB, Status: Passed - Codetree

+++ 251118/길이 단위 환산하기/convert-length-units.py
@@ -1,2 +1,2 @@
-a = int(input())
+a = float(input())
 print(f"{a*30.48:.1f}")
